% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/boot.R
\name{boot}
\alias{boot}
\title{\code{boot}}
\usage{
boot(i, G, formula, data, family, beta_model, cluster_se, indep_vars, k,
  clust_list, clust_length, stratify, nonparametric)
}
\arguments{
\item{i}{fake argument to pass to mclapply}

\item{G}{number of groups/clusters}

\item{formula}{model formula}

\item{data}{data.fram to use in analysis}

\item{family}{character string, which model to use (binomial or gaussian)}

\item{beta_model}{numeric vector, estimated coefs from model}

\item{cluster_se}{logical, use clustered standard errors?}

\item{indep_vars}{character vector of names of independent variables}

\item{k}{number of regressors}

\item{clust_list}{list of integer vectors with row indices of cluster members}

\item{clust_length}{integer vector of num of observations in each cluster}

\item{stratify}{logical, further resample within clusters?}

\item{nonparametric}{logical, conduct nonparametric bootstrap?}
}
\value{
numeric vector of estimated coefficients (if nonparametric) or
     p-values (if parametric)
}
\description{
Calculate one clustered bootstrap replicate (for internal use only)
}

