% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cl.R
\name{cl}
\alias{cl}
\title{\code{cl}}
\usage{
cl(data, model, cluster)
}
\arguments{
\item{data}{data.frame to use}

\item{model}{fitted model from glm}

\item{cluster}{integer vector of group indices}
}
\value{
output from coeftest
}
\description{
Calculate clustered standard errors
}

