% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cluster_bs.R
\name{cluster_bs}
\alias{cluster_bs}
\title{\code{cluster_bs}}
\usage{
cluster_bs(model, data, cluster, ci_level = 0.95, boot_reps = 1000L,
  stratify = FALSE, cluster_se = TRUE, report = TRUE,
  nonparametric = TRUE, parallel = TRUE)
}
\arguments{
\item{model}{fitted model}

\item{data}{data.frame to use in estimation}

\item{cluster}{RHS formula to identify clusters}

\item{ci_level}{numeric in (0, 1), level for inference}

\item{boot_reps}{integer, number of bootstrap replicates}

\item{stratify}{logical, should bootstrapping within clusters be done?}

\item{cluster_se}{logical, use SEs clustered at cluster?}

\item{report}{logical, print to screen?}

\item{nonparametric}{logical, do nonparametric bootstrap? if FALSE do
    parametric}

\item{parallel}{logical, use multiple cores?}
}
\value{
list with p values, confidence intervals, and SEs
}
\description{
Calculate bootstrapped clustered standard errors
}

