% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/estimate.R
\name{estimate}
\alias{estimate}
\title{\code{estimate}}
\usage{
estimate(net, boot_reps = 10, data = NULL, exogs = NULL,
  incl_contagion = TRUE, incl_conflict = TRUE, incl_vote_intent = TRUE,
  contagion_varnames = "mvint", conflict_varnames = c("ideomd", "ideosd0",
  "party.rhhi"), tie = "both", model = "logit", subs = NA,
  parallel = TRUE, incl_chapter_FEs = TRUE)
}
\arguments{
\item{net}{character, name of network type}

\item{boot_reps}{integer, number of bootstrap replicates}

\item{data}{data.frame to use in model}

\item{exogs}{character vector of exogenous variable names to include in model}

\item{incl_contagion}{logical, whether to include contagion analysis}

\item{incl_conflict}{logical, whether to include conflict analysis}

\item{incl_vote_intent}{logical, whether to include voting intent}

\item{contagion_varnames}{character vector of contagion variable names}

\item{conflict_varnames}{character vector of conflict variable names}

\item{tie}{characater, argument to pass to make_vars, which tie type to use
    (both, in, or out)}

\item{model}{character, link function name (logit or gaussian)}

\item{subs}{vector describing how to subset, either NA for no subsetting or
    a pair of values, one with column name to subset by and a second with
    value to subset with}

\item{parallel}{logical, use multiple cores? overridden if b < 100}

\item{incl_chapter_FEs}{logical, include chapter fixed effects?}
}
\value{
object of classes cblest and est
}
\description{
Estimate peer effects model
}

