% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/format_object.R
\name{format_object}
\alias{format_object}
\title{\code{format_object}}
\usage{
format_object(object, coef_type = "as.is", SE = FALSE)
}
\arguments{
\item{object}{est object}

\item{coef_type}{character, type of coefficients to show (as.is, OR, prob, or
    perc.change)}

\item{SE}{logical, print SE if TRUE, CI if FALSE}
}
\value{
character vector
}
\description{
Format the ouput from a single model object (for internal use)
}

