% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/format_output.R
\name{format_output}
\alias{format_output}
\title{\code{format_output}}
\usage{
format_output(..., full = TRUE, coef_type = "as.is")
}
\arguments{
\item{...}{est object(s) to format for output}

\item{full}{logical, report extra controls}

\item{coef_type}{character, type of coefficients to show (as.is, OR, prob, or
     perc.change)}
}
\value{
character array
}
\description{
Format the ouput from one object or a list of objects
}

