% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/make_similarity_matrix.R
\name{make_similarity_matrix}
\alias{make_similarity_matrix}
\title{\code{make_similarity_matrix}}
\usage{
make_similarity_matrix(attrx)
}
\arguments{
\item{attrx}{vector of attributes for nodes, length n}
}
\value{
n x n array of (binary) similarity on attrx
}
\description{
Produce adjacency matrix for similarity on input argument
}

