% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/make_vars.R
\name{make_vars}
\alias{make_vars}
\title{\code{make_vars}}
\usage{
make_vars(net, tie = "both", contagion_varnames = "mvint",
  conflict_varnames = c("ideomd", "ideosd0", "party.rhhi"))
}
\arguments{
\item{net}{character vector, type of network}

\item{tie}{character, type of tie (both, in, out)}

\item{contagion_varnames}{character vector of variables for peer effects}

\item{conflict_varnames}{character vector of variables for peer effects}
}
\value{
list of contagion and conflict variables
}
\description{
Generates the contagion and conflict peer variables (for internal use)
}

