% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/marg_gen.R
\name{marg_gen}
\alias{marg_gen}
\title{\code{marg_gen}}
\usage{
marg_gen(object, peer_int = 0.2, peer_ref = c(1, 5), peer_varname = NULL,
  cond_varname = NULL, cond_value = NULL)
}
\arguments{
\item{object}{est object}

\item{peer_int}{comparison value}

\item{peer_ref}{reference value}

\item{peer_varname}{peer effect variable name}

\item{cond_varname}{conditional effect variable name}

\item{cond_value}{conditional effect value}
}
\value{
data.frame summarizing marginal effects
}
\description{
Calculate marginal effects
}

