% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/merge_mice.R
\name{merge_mice}
\alias{merge_mice}
\title{\code{merge_mice}}
\usage{
merge_mice(imp)
}
\arguments{
\item{imp}{list of mids objects}
}
\value{
mids object
}
\description{
taken from http://stackoverflow.com/questions/24040280,
 (with) minor bug fixes
}

