% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mi_estimate.R
\name{mi_estimate}
\alias{mi_estimate}
\title{\code{mi_estimate}}
\usage{
mi_estimate(net, m = 10L, data = NULL, nom = c("Female", "Race_white",
  "Race_black", "Rep", "Dem", "Ind"), ord = c("SchYear", "Ideology",
  "Ideo_strength", "Interest_pol"), incl_imputations = FALSE,
  contagion_varnames = "mvint", conflict_varnames = c("ideomd", "ideosd0",
  "party.rhhi"), incl_contagion = TRUE, incl_conflict = FALSE,
  incl_vote_intent = TRUE, fast = FALSE, tie = "both", model = "logit",
  subs = NA, incl_chapter_FEs = TRUE, parallel = TRUE)
}
\arguments{
\item{net}{character, name of network type}

\item{m}{integer number of imputations to perform}

\item{data}{data.frame to use in model}

\item{nom}{character vector of names of variables with nominal data}

\item{ord}{character vector of names of variables with ordered data}

\item{incl_imputations}{logical, store imputations?}

\item{contagion_varnames}{character vector of contagion variable names}

\item{conflict_varnames}{character vector of conflict variable names}

\item{incl_contagion}{logical, whether to include contagion analysis}

\item{incl_conflict}{logical, whether to include conflict analysis}

\item{incl_vote_intent}{logical, whether to include voting intent}

\item{fast}{logical, TODO fill this out}

\item{tie}{characater, argument to pass to make_vars, which tie type to use
    (both, in, or out)}

\item{model}{character, link function name (logit or gaussian)}

\item{subs}{vector describing how to subset, either NA for no subsetting or
    a pair of values, one with column name to subset by and a second with
    value to subset with}

\item{incl_chapter_FEs}{logical, include chapter fixed effects?}

\item{parallel}{logical, use multiple cores? overridden if m <= 10}
}
\value{
object of class miest
}
\description{
Do multiple imputation and estimation
}

