% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/peer_calc.R
\name{peer_calc}
\alias{peer_calc}
\title{\code{peer_calc}}
\usage{
peer_calc(varname, attrx, stat = "mean")
}
\arguments{
\item{varname}{character, variable name from attrx}

\item{attrx}{data.frame of node-levels variables}

\item{stat}{"mean" or "div", which peer effects to calculate}
}
\value{
data.frame, results of a call to either peer_mean or peer_div
}
\description{
Generate the various conflict & contagion variables
}

