% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/peer_div.R
\name{peer_div}
\alias{peer_div}
\title{\code{peer_div}}
\usage{
peer_div(netx, attrx, varname = "ideomd", tie = "both")
}
\arguments{
\item{netx}{adjacency matrix of ties}

\item{attrx}{data.frame of node-levels variables}

\item{varname}{character, variable name to output}

\item{tie}{character, tie type (both, in, out)}
}
\value{
data.frame
}
\description{
Calculate summary peer diversity values
}

