% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/peer_mean.R
\name{peer_mean}
\alias{peer_mean}
\title{\code{peer_mean}}
\usage{
peer_mean(netx, attrx, varname = "vint", tie = "both")
}
\arguments{
\item{netx}{adjacency matrix of ties}

\item{attrx}{data.frame of node-levels variables}

\item{varname}{character, variable name from attrx}

\item{tie}{character, type type (both, in, out)}
}
\value{
data.frame
}
\description{
calculate peer summary
}

