% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_marginal.R
\name{plot_marginal}
\alias{plot_marginal}
\title{\code{plot_marginal}}
\usage{
plot_marginal(meff)
}
\arguments{
\item{meff}{data.frame of marginal effects}
}
\value{
object of class ggplot
}
\description{
Draw a marginal effects plot
}

