% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pool_r2.R
\name{pool_r2}
\alias{pool_r2}
\title{\code{pool_r2}}
\usage{
pool_r2(object, adjusted = FALSE)
}
\arguments{
\item{object}{object of class miest (mira)}

\item{adjusted}{use adjusted R2?}
}
\value{
data.frame of results
}
\description{
Compute pooled R squared
}

