% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pretty_var_names.R
\name{pretty_var_names}
\alias{pretty_var_names}
\title{\code{pretty_var_names}}
\usage{
pretty_var_names(vars)
}
\arguments{
\item{vars}{integer vector of which variable names to return}
}
\value{
character vector
}
\description{
Makes variable names for publication
}

