% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/print.est.R
\name{print.est}
\alias{print.est}
\title{\code{print.est}}
\usage{
\method{print}{est}(x, ...)
}
\arguments{
\item{x}{a fitted model}

\item{...}{don't use this}
}
\value{
argument x, invisibly
}
\description{
print output from estimate or mi_estimate
}

