% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/printmat.R
\name{printmat}
\alias{printmat}
\title{\code{printmat}}
\usage{
printmat(m)
}
\arguments{
\item{m}{matrix}
}
\value{
results of write.table
}
\description{
Format a matrix of results (for internal use)
}

