% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pseudoR2.R
\name{pseudoR2}
\alias{pseudoR2}
\title{\code{pseudoR2}}
\usage{
pseudoR2(object)
}
\arguments{
\item{object}{object of class est}
}
\value{
vector of summary statistics
}
\description{
report various model fit measures (adapted from pR2 in pscl package)
}

