% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/strip0.R
\name{strip0}
\alias{strip0}
\title{\code{strip0}}
\usage{
strip0(val, digits = 2L)
}
\arguments{
\item{val}{numeric (or character with numbers) to reformat}

\item{digits}{integer, number of significant digits}
}
\value{
character
}
\description{
remove leading zeros, set significant digits
}

