% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/summary.est.R
\name{summary.est}
\alias{summary.est}
\title{\code{summary.est}}
\usage{
\method{summary}{est}(object, full = FALSE, quietly = FALSE,
  coef_type = "as.is", ...)
}
\arguments{
\item{object}{object of class est}

\item{full}{logical, present all coefficients}

\item{quietly}{logical, print results to screen}

\item{coef_type}{character, type of coefficients to show (as.is, OR, prob, or
    perc.change)}

\item{...}{don't use this}
}
\value{
data.frame, invisibly
}
\description{
summarize results of estimation
}

